Migrate Subscription Manager Magic Link page to Inertia
- Extract magic_link and send_magic_link actions from SubscriptionsController
  to a dedicated Subscriptions::MagicLinksController (new/create)
- Convert ERB view to Inertia page component (Subscriptions/MagicLink)
- Remove old server-component, webpack pack, and SSR registration
- Move controller tests to new spec file with Inertia assertions
- Preserve existing route paths and names for backward compatibility

Closes #3072
