+++ app/javascript/ssr.ts
@@ -21,7 +21,6 @@ import Profile from "$app/components/server-components/Profile";
 import ProfileProductPage from "$app/components/server-components/Profile/ProductPage";
 import SubscribePage from "$app/components/server-components/SubscribePage";
 import SubscriptionManager from "$app/components/server-components/SubscriptionManager";
-import SubscriptionManagerMagicLink from "$app/components/server-components/SubscriptionManagerMagicLink";
 import SupportHeader from "$app/components/server-components/support/Header";
 import TaxesCollectionModal from "$app/components/server-components/TaxesCollectionModal";
 import VideoStreamPlayer from "$app/components/server-components/VideoStreamPlayer";
@@ -50,7 +49,6 @@ ReactOnRails.register({
   ProfileProductPage,
   SubscribePage,
   SubscriptionManager,
-  SubscriptionManagerMagicLink,
   TaxesCollectionModal,
   VideoStreamPlayer,
 });
