--- app/views/subscriptions/magic_link.html.erb
@@ -1,4 +1,0 @@
-<%= render("shared/recaptcha_script") %>
-<%= load_pack("subscription_magic_link") %>
-
-<%= react_component "SubscriptionManagerMagicLink", props: @react_component_props, prerender: true %>
